gradient and gradient descend 2d working

+++ functions/rk4_2D.h
@@ -2,6 +2,6 @@
 #define RK4_2D_H
 #include <stdbool.h>
 
-void gradient_descend_rk4(double* point, double *matriz, int filas, int columnas, double *new_point, int step);
+void gradient_descend_rk4(double* point, double *matriz, int filas, int columnas, double *new_point, double step);
 
 #endif
